requirements.txt
pandas==1.0.0
numpy==1.18.1
altair==4.0.1
streamlit==0.56.0
python-dateutil==2.8.0
pyyaml==5.3
cachetools==4.0.0